Move Measures::BulkBaseController out of Measures namespace as it is generic
