+++ app/controllers/additional_codes/bulk_items_controller.rb
@@ -1,5 +1,5 @@
 module AdditionalCodes
-  class BulkItemsController < Measures::BulksBaseController
+  class BulkItemsController < BulksBaseController
     before_action :require_to_be_workbasket_owner!, only: [
       :remove_items
     ]
