+++ app/controllers/measures/bulks_controller.rb
@@ -1,5 +1,5 @@
 module Measures
-  class BulksController < Measures::BulksBaseController
+  class BulksController < BulksBaseController
     include ::SearchCacheHelpers
 
     before_action :require_to_be_workbasket_owner!, only: %i[
